Release 0.1.0
Since skolmaten.se recently made alot of changes to their site and its APIs, this package needs some changes too (since it broke).

At skolmaten.se, it says that you need an API-key to query their API. Thats not the case, its (for now) only required to set the referer of the http request to their domain.

A quick summary of changes:
- Adds nix flake configurations
- Replaces old requirements.txt with poetry
- Add tests
- Add Github workflows for tests and pypi publish
- Pre-commit
- Updated readme